# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WatchStore Api List",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/products_by_category/{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "816c7ec9f6cc34f2ebf4dbd505eebad6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/products_by_brand{id}": {
            "get": {
                "tags": [
                    "Products Page"
                ],
                "description": "get products data by category id",
                "operationId": "31390a9d4c65629eb9540440e246c0ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/product_details/{id}": {
            "get": {
                "tags": [
                    "Product Details"
                ],
                "description": "get product details data by product id",
                "operationId": "be2ea286eec642c3ff6b24853c417d78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}